chore: add test of output option
