+++ __tests__/index.spec.ts
@@ -1,6 +1,7 @@
 import fs from 'fs'
 import path from 'path'
 import { version } from '../package.json'
+import { projects } from '../projects/projects'
 import build, { resetCache } from '../src/buildTemplate'
 import { run } from '../src/cli'
 import getConfig from '../src/getConfig'
@@ -18,20 +19,20 @@ describe('cli test', () => {
   })
 
   test('main', async () => {
-    for (const dir of fs.readdirSync('projects')) {
-      if (dir.endsWith('.ts')) continue
+    const pjs = projects.flatMap(project => [{ ...project, output: undefined }, project])
 
+    for (const project of pjs) {
       resetCache()
 
-      const workingDir = path.join(process.cwd(), 'projects', dir)
+      const workingDir = path.join(process.cwd(), 'projects', project.dir)
       const { type, input, staticDir, output, trailingSlash } = await getConfig(
-        dir !== 'nuxtjs-no-slash',
-        undefined,
+        project.options.includes('-s') || project.options.includes('--enableStatic'),
+        project.output && path.join(workingDir, project.output),
         workingDir
       )
 
       const result = fs.readFileSync(`${output}/$path.ts`, 'utf8')
-      const basepath = /-basepath$/.test(dir) ? '/foo/bar' : undefined
+      const basepath = /-basepath$/.test(project.dir) ? '/foo/bar' : undefined
       const { filePath, text } = build({ type, input, staticDir, output, trailingSlash, basepath })
 
       expect(filePath).toBe(`${output}/$path.ts`)
